# csharp/api/api_documentation.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "src | v1",
    "version": "1.0.0"
  },
  "paths": {
    "/hexagon/get-countries-with-cities": {
      "get": {
        "tags": [
          "Hexagon"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/hexagon/get-hexagons-from-city": {
      "get": {
        "tags": [
          "Hexagon"
        ],
        "parameters": [
          {
            "name": "city",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/hexagon/get-pois-from-hexagon": {
      "get": {
        "tags": [
          "Hexagon"
        ],
        "parameters": [
          {
            "name": "hexagonId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/hexagon/post-location-batch": {
      "post": {
        "tags": [
          "Hexagon"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostLocationBatchDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PostLocationBatchDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PostLocationBatchDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/hexagon/get-hexagon-progresses": {
      "get": {
        "tags": [
          "Hexagon"
        ],
        "parameters": [
          {
            "name": "city",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/user/validate-login-token": {
      "post": {
        "tags": [
          "User"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            },
            "text/json": {
              "schema": {
                "type": "string"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/user/get-logged-user": {
      "post": {
        "tags": [
          "User"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/user/get-user-statistics": {
      "get": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "name": "city",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/user/delete-user-account": {
      "post": {
        "tags": [
          "User"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/version/get-current-version": {
      "get": {
        "tags": [
          "Version"
        ],
        "parameters": [
          {
            "name": "key",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "LocationDto": {
        "type": "object",
        "properties": {
          "latitude": {
            "type": "number",
            "format": "double"
          },
          "longitude": {
            "type": "number",
            "format": "double"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "PostLocationBatchDto": {
        "required": [
          "locations"
        ],
        "type": "object",
        "properties": {
          "locations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LocationDto"
            }
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Hexagon"
    },
    {
      "name": "User"
    },
    {
      "name": "Version"
    }
  ]
}